pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}

[tool.isort]
multi_line_output = 3
lines_after_imports = 2
